Move `SplitScroll` into `egui_table`
