+++ egui_table/src/lib.rs
@@ -1,1 +1,5 @@
-//! Example of a Rust library.
+//! Table viewer for [egui](https://www.egui.rs/).
+
+pub mod split_scroll;
+
+pub use split_scroll::{SplitScroll, SplitScrollDelegate};
